Renamed "device" to "node"
Previously we were calling target computer as "device" which is more
familiar name for IoT computers, "devices".
But in future we want to bring the attached devices (sensors,
usb devices, harddrives) into the model so "Device" name
were conflicting with it.
Renamed Device concept to Node as it is in Kubernetes so we can
use Device to describe attached devices.